refactor: move helpers to util
