--- src/helpers/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-module.exports = {
-  dataHelper: require('./data'),
-  requestHelper: require('./request'),
-  timeHelper: require('./time')
-}
