+++ src/services/exile.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { ConflictError, ForbiddenError, NotFoundError } = require('../errors')
-const { inRange } = require('../helpers').dataHelper
+const { inRange } = require('../util').util
 const { Exile } = require('../models')
 
 const applicationConfig = require('../../config/application')
