+++ src/util/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = {
+  requestUtil: require('./request'),
+  timeUtil: require('./time'),
+  util: require('./util')
+}
